+++ CHANGELOG.md
@@ -49,6 +49,12 @@
   - <https://github.com/vivliostyle/vivliostyle.js/pull/116>
 - Fix bug that rules above footnotes disappear
   - <https://github.com/vivliostyle/vivliostyle.js/pull/118>
+- Allow an EPUB directory URL not followed by a slash
+  - <https://github.com/vivliostyle/vivliostyle.js/pull/120>
+- Change initial values of `orphans` and `widows` to 2
+  - <https://github.com/vivliostyle/vivliostyle.js/issues/30>
+- Allow page/column break inside tables
+  - <https://github.com/vivliostyle/vivliostyle.js/issues/101>
 
 ## [0.2.0](https://github.com/vivliostyle/vivliostyle.js/releases/tag/0.2.0) - 2015-09-16
 Beta release.